03_Langchain_models/02.Chat_Models/2.chatmodel_anthropic.py
from langchain_anthropic import ChatAnthropic
from dotenv import load_dotenv

load_dotenv()

model = ChatAnthropic(model="claude-2", temperature=0.7)

result = model.invoke('Explain the theory of relativity in simple terms.')
print(result)